requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.4.2
cryptography==41.0.4